Add files for tracking
