+++ src/main/java/oscar/Oscar.java
@@ -1,10 +1,20 @@
+package oscar;
+
+import oscar.command.Command;
+import oscar.essential.Parser;
+import oscar.essential.Storage;
+import oscar.essential.TaskList;
+import oscar.essential.Ui;
+
+import oscar.exception.OscarException;
+
 import java.util.Scanner;
 
 /**
  * Chatbot named Oscar that can respond to user input.
  */
 public class Oscar {
-    static final String FILEPATH = "./data/talist";
+    static final String FILEPATH = "./data/tasklist";
 
     private final Storage storage;
     private TaskList tasks;
