+++ src/main/java/oscar/command/Command.java
@@ -1,3 +1,11 @@
+package oscar.command;
+
+import oscar.essential.Storage;
+import oscar.essential.TaskList;
+import oscar.essential.Ui;
+
+import oscar.exception.OscarException;
+
 import java.time.format.DateTimeFormatter;
 
 /**
