+++ src/main/java/oscar/command/DeadlineCommand.java
@@ -1,3 +1,14 @@
+package oscar.command;
+
+import oscar.essential.Storage;
+import oscar.essential.TaskList;
+import oscar.essential.Ui;
+
+import oscar.exception.OscarException;
+
+import oscar.task.Deadline;
+import oscar.task.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
