+++ src/main/java/oscar/command/DeleteCommand.java
@@ -1,20 +1,28 @@
+package oscar.command;
+
+import oscar.essential.Storage;
+import oscar.essential.TaskList;
+import oscar.essential.Ui;
+
+import oscar.exception.OscarException;
+
 /**
- * Command to mark a task as done.
+ * Command to delete a task from the task list.
  */
-public class MarkCommand extends Command {
+public class DeleteCommand extends Command {
     private final String details;
 
     /**
-     * Instantiates a mark command.
-     * @param details Task number to be marked.
+     * Instantiates a delete command.
+     * @param details Task number to be deleted.
      */
-    public MarkCommand(String details) {
+    public DeleteCommand(String details) {
         super(false);
         this.details = details;
     }
 
     /**
-     * Marks a task as done using the task number.
+     * Deletes a task using the task number.
      * @param tasks ArrayList of tasks.
      * @param ui User interaction handler.
      * @param storage File loading and saving handler.
@@ -23,9 +31,10 @@ public class MarkCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws OscarException {
         int index = validateInt(details, tasks);
-        String currentTask = tasks.mark(index);
+        String currentTask = tasks.delete(index);
         storage.save(tasks);
-        System.out.println("Nice! Oscar has marked this task as done:\n");
+        System.out.println("Oscar has removed this task:\n");
         System.out.println(currentTask + "\n");
+        tasks.listCount();
     }
 }
