+++ src/main/java/oscar/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package oscar.command;
+
+import oscar.essential.Storage;
+import oscar.essential.TaskList;
+import oscar.essential.Ui;
+
 /**
  * Command to terminate Oscar.
  */
