+++ src/main/java/oscar/task/Event.java
@@ -1,3 +1,5 @@
+package oscar.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
